[PDW-6] chore: 컴포넌트 별 파일 분리
