+++ src/components/officeBooking/BookingOfficeInfo.js
@@ -1,0 +1,45 @@
+import styled from "styled-components"
+
+
+export const OfficeInfoContainer = styled.div`
+    width: 94%;
+    height: 299px;
+    border-radius: 12px;
+    background: #FFF;
+    margin: 3%;
+    border: 1px solid #E6E6E6;
+`
+
+export const OfficeImgsContainer = styled.div`
+    width: 390px;
+    height: 255px;
+    margin-top: 22px;
+    margin-left: 35px;
+    background: #2A3042;
+    float: left;
+`
+
+export const OfficeDetailContainer = styled.div`
+    float: left;
+    margin-top: 10px;
+    margin-left: 15px;
+`
+
+export const OfficeDetailCapsuleContainer = styled.div`
+    width: 50px;
+`
+
+export const OfficeDetailTextContainer = styled.div`
+    width: 800px;
+    height: 150px;
+    float: left;
+`
+
+export const OfficeDetailText = styled.text`
+    font-family: NanumSquare_ac;
+    font-size: 20px;
+    font-weight: 400;
+    line-height: 25px;
+    letter-spacing: 0em;
+    text-align: left;
+`
