+++ src/components/officeBooking/BookingPurpose.js
@@ -1,0 +1,16 @@
+import styled from "styled-components"
+
+export const BookingPurposeContainer = styled.div`
+    width: 94%;
+    height: 100px;
+    margin: 3%;
+`
+
+export const BookingPurposeTextFieldContainer = styled.div`
+    margin-left: 20px;
+    float: left;
+`
+
+export const BookingCapsuleContainer = styled.div`
+    float: left;
+`
