+++ src/components/officeBooking/SubTitleBar.js
@@ -1,0 +1,34 @@
+import styled from "styled-components"
+
+export const SubTitleContainer = styled.div`
+    width: 100%;
+    height: 64px;
+    border-radius: 12px;
+    background: #2A3042;
+`
+
+export const SubTitleTextContainer = styled.div`
+    margin-left: 10px;
+    margin-top: 15px;
+    float: left;
+`
+
+export const SelectedSubTitleText = styled.text`
+    font-family: NanumSquare_ac;
+    font-size: 24px;
+    font-weight: 700;
+    line-height: 27px;
+    letter-spacing: 0em;
+    text-align: left;
+    color: #FFFFFF;
+`
+
+export const UnselectedSubTitleText = styled.text`
+    font-family: NanumSquare_ac;
+    font-size: 20px;
+    font-weight: 400;
+    line-height: 22px;
+    letter-spacing: 0em;
+    text-align: left;
+    color: #FFFFFF;
+`
