+++ src/pages/booking/officeBooking/OfficeInfo.js
@@ -2,14 +2,12 @@ import React from 'react';
 import styled from "styled-components"
 import Capsule from 'components/capsule/Capsule';
 
-var selectedCheckList = [
-    false, true, false, false, false, false, false, false, false, false, false, false, 
-    false, true, false, false, false, false, false, false, false, false, false, false, 
-];
+import {SubTitleContainer, SubTitleTextContainer, SelectedSubTitleText, UnselectedSubTitleText} from 'components/officeBooking/SubTitleBar';
+import {OfficeInfoContainer, OfficeImgsContainer, OfficeDetailContainer, OfficeDetailCapsuleContainer, OfficeDetailTextContainer, OfficeDetailText } from 'components/officeBooking/BookingOfficeInfo';
+import {BookingContentContainer, BookingTimeContainer, renderBookingTimeBar} from 'components/officeBooking/BookingTimeBar';
+import {BookingPurposeContainer, BookingCapsuleContainer, BookingPurposeTextFieldContainer} from 'components/officeBooking/BookingPurpose';
+import {RequestBookingButton, requestBookingOffice} from 'components/officeBooking/BookingRequest';
 
-function isSelectedTime(time) {
-    return selectedCheckList[time];
-}
 
 export const Container = styled.div`
     width: 87%;
@@ -36,219 +34,6 @@ export const ContentContainer = styled.div`
     margin-top: 20px;
 `
 
-export const SubTitleContainer = styled.div`
-    width: 100%;
-    height: 64px;
-    border-radius: 12px;
-    background: #2A3042;
-`
-
-export const SubTitleTextContainer = styled.div`
-    margin-left: 10px;
-    margin-top: 15px;
-    float: left;
-`
-
-export const SelectedSubTitleText = styled.text`
-    font-family: NanumSquare_ac;
-    font-size: 24px;
-    font-weight: 700;
-    line-height: 27px;
-    letter-spacing: 0em;
-    text-align: left;
-    color: #FFFFFF;
-`
-
-export const UnselectedSubTitleText = styled.text`
-    font-family: NanumSquare_ac;
-    font-size: 20px;
-    font-weight: 400;
-    line-height: 22px;
-    letter-spacing: 0em;
-    text-align: left;
-    color: #FFFFFF;
-`
-
-export const OfficeInfoContainer = styled.div`
-    width: 94%;
-    height: 299px;
-    border-radius: 12px;
-    background: #FFF;
-    margin: 3%;
-    border: 1px solid #E6E6E6;
-`
-
-export const OfficeImgsContainer = styled.div`
-    width: 390px;
-    height: 255px;
-    margin-top: 22px;
-    margin-left: 35px;
-    background: #2A3042;
-    float: left;
-`
-
-export const OfficeDetailContainer = styled.div`
-    float: left;
-    margin-top: 10px;
-    margin-left: 15px;
-`
-
-export const OfficeDetailCapsuleContainer = styled.div`
-    width: 50px;
-`
-export const BookingCapsuleContainer = styled.div`
-    float: left;
-`
-
-export const OfficeDetailText = styled.text`
-    font-family: NanumSquare_ac;
-    font-size: 20px;
-    font-weight: 400;
-    line-height: 25px;
-    letter-spacing: 0em;
-    text-align: left;
-`
-
-export const OfficeDetailTextContainer = styled.div`
-    width: 800px;
-    height: 150px;
-    float: left;
-`
-
-export const BookingContentContainer = styled.div`
-    margin-left: 35px;
-`
-
-export const BookingPurposeContainer = styled.div`
-    width: 94%;
-    height: 100px;
-    margin: 3%;
-`
-
-export const BookingPurposeTextFieldContainer = styled.div`
-    margin-left: 20px;
-    float: left;
-`
-
-export const contentText = styled.text`
-    font-family: NanumSquare_ac;
-    font-size: 25px;
-    font-weight: 400;
-    line-height: 16px;
-    letter-spacing: 0em;
-    text-align: left;
-`
-
-export const RequestBookingButton = styled.button`
-    borger: none;
-    padding: 5px 10px;
-    background-color: #A263DE;
-    color: #FFF;
-    width: 82px;
-    height: 38px;
-    border: none;
-    border-radius: 20px;
-
-    font-family: NanumSquare_ac;
-    font-size: 20px;
-    font-weight: 400;
-    line-height: 16px;
-    letter-spacing: 0em;
-    text-align: center;
-`
-
-export const BookingTimeContainer = styled.div`
-    width: 94%;
-    hegith: 60px;
-    margin-left: 45px;
-`
-
-export const FirstBookingTimeButton = styled.button`
-    width: 47px;
-    height: 30px;
-    background-color: ${color => color};
-    margin-left: 2px;
-    margin-right: 3px;
-    border: none;
-    border-radius: 15px 0px 0px 15px; 
-    onClick = clickTimeBarItem(0);
-`
-
-export const LastBookingTimeButton = styled.button`
-    width: 47px;
-    height: 30px;
-    background-color: ${color => color};
-    margin-left: 3px;
-    margin-right: 1px;
-    border: none;
-    border-radius:  0px 15px 15px 0px; 
-    onClick = clickTimeBarItem(0);
-`
-
-export const BookingTimeButton = styled.button`
-    width: 47px;
-    height: 30px;
-    background-color: ${color => color};
-    border: none;
-    onClick = clickTimeBarItem(0);
-`
-
-export const TimeButtonContainer = styled.div`
-    width: 47px;
-    float: left;
-    margin-left: 2px;
-    margin-right: 3px;
-    margin-bottom: 30px;
-`
-
-export const StartTimeTextContainer = styled.div`
-    float: left;
-`
-
-export const EndTimeTextContainer = styled.div`
-    float: right;
-`
-
-
-function renderBookingTimeButtonItem(index) {
-    if (index == 0) {
-        return <TimeButtonContainer>
-            <FirstBookingTimeButton color={isSelectedTime(index) ? '#81c147' : '#D3D3D3'} onClick={clickTimeBarItem(index)}/>
-            <StartTimeTextContainer>{index}</StartTimeTextContainer>
-        </TimeButtonContainer>
-    } else if (index == 23) {
-        return <TimeButtonContainer>
-            <LastBookingTimeButton color={isSelectedTime(index) ? '#81c147' : '#D3D3D3'} onClick={clickTimeBarItem(index)}/>
-            <StartTimeTextContainer>{index}</StartTimeTextContainer>
-            <EndTimeTextContainer>{index+1}</EndTimeTextContainer>
-        </TimeButtonContainer>
-    } else {
-        return <TimeButtonContainer>
-            <BookingTimeButton color={isSelectedTime(index) ? '#81c147' : '#D3D3D3'} onClick={clickTimeBarItem(index)}/>
-            <StartTimeTextContainer>{index}</StartTimeTextContainer>
-        </TimeButtonContainer>
-    }
-}
-
-function clickTimeBarItem(index) {
-    console.log('바뀌기 전 -> ', selectedCheckList[index]);
-    selectedCheckList[index] = !selectedCheckList[index];
-    console.log('바뀐 후 -> ', selectedCheckList[index]);
-}
-
-function renderBookingTimeBar() {
-    var items = [];
-    for (var i = 0; i < 24; i++) {
-        items.push(renderBookingTimeButtonItem(i));
-    }
-    return items;
-}
-
-function requestBookingOffice() {
-    var bookingPurpose = document.getElementById("bookingPurpose").value;
-    console.log("예약목적 : ", bookingPurpose);
-    alert("회의실을 예약하겠습니까?");
-}
 
 function OfficeBooking() {
     return (
